Refactor sender/exporter
